Safari-611.1.10/LayoutTests/media/media-controls-timeline-updates-expected.txt

This tests that media controls timeline updates during playback.

This test only runs in DRT!


EVENT(canplaythrough)
EXPECTED (timeLineValue() == '0') OK
EVENT(playing)
EVENT(pause)
TEST(video.currentTime > 0) OK
TEST(timeLineValue() > 0) OK
END OF TEST
